+++ diplomacy/map_parser/vector/transform.py
@@ -3,9 +3,21 @@ from abc import abstractmethod
 from xml.etree.ElementTree import Element
 import numpy as np
 
+
+# TODO: Refactor all of this into one Transform (Empty and Translation can be trivally converted into Matrix)
+
 class Transform:
     def __init__(self, element: Element):
-        self.transform_string: str | None = element.get("transform", None)
+        if element is not None:
+            self.transform_string: str | None = element.get("transform", None)
+        else:
+            self.transform_string = None
+        self.x_dx = 1
+        self.y_dy = 1
+        self.x_dy = 0
+        self.y_dx = 0
+        self.x_c = 0
+        self.y_c = 0
 
     @abstractmethod
     def transform(self, point: tuple[float, float]) -> tuple[float, float]:
@@ -19,6 +31,12 @@ class Transform:
 class EmptyTransform(Transform):
     def __init__(self, element: Element):
         super().__init__(element)
+        self.x_dx = 1
+        self.y_dy = 1
+        self.x_dy = 0
+        self.y_dx = 0
+        self.x_c = 0
+        self.y_c = 0
 
     def transform(self, point: tuple[float, float]) -> tuple[float, float]:
         return point
@@ -32,33 +50,41 @@ class EmptyTransform(Transform):
 
 class Translation(Transform):
     def __init__(self, element: Element, data=None):
+        super().__init__(element)
         if data != None:
             # just set it to this value
-            print(type(data[0]))
             assert(not isinstance(data, str))
-            self.x = data[0]
-            self.y = data[1]
+            self.x_c = data[0]
+            self.y_c = data[1]
             return
 
-        super().__init__(element)
-        self.x, self.y = 0, 0
+        self.x_c, self.y_c = 0, 0
 
         if self.transform_string:
             translation_match = re.search("^\\s*translate\\((.*),(.*)\\)\\s*", self.transform_string)
             if translation_match:
-                self.x = float(translation_match.group(1))
-                self.y = float(translation_match.group(2))
+                self.x_c = float(translation_match.group(1))
+                self.y_c = float(translation_match.group(2))
             else:
                 raise RuntimeError("Translation not found")
+        self.x_dx = 1
+        self.y_dy = 1
+        self.x_dy = 0
+        self.y_dx = 0
 
     def transform(self, point: tuple[float, float]) -> tuple[float, float]:
-        return point[0] + self.x, point[1] + self.y
+        return point[0] + self.x_c, point[1] + self.y_c
 
     def inverse_transform(self, point):
-        return point[0] - self.x, point[1] - self.y
+        return point[0] - self.x_c, point[1] - self.y_c
 
     def __str__(self):
-        return f"translate({self.x},{self.y})"
+        return f"translate({self.x_c},{self.y_c})"
+
+
+
+    def __mul__(self, other):
+        selfm = np.array([[self.x_dx, self.y_dx, 0], [self.x_dy, self.y_dy, 0], [self.x_c, self.y_c, 1]])
 
 class MatrixTransform(Transform):
     def __init__(self, element: Element):
