+++ requirements.txt
@@ -1,7 +1,7 @@
 numpy
 scipy
-pydip
 shapely
 lark
 lxml
 discord
+black
